# docs/swagger.yaml
openapi: 3.0.3
info:
  title: 'Personal Expense Tracker API with Chatbot Analyst'
  description: |
    A RESTful API for tracking personal expenses with integrated chatbot analysis.

    ## Authentication
    The API uses session-based authentication with HTTP-only cookies.
    - Register or login to receive a `session_token` cookie (7-day expiry)
    - Session tokens are stored server-side as SHA-256 hashes and are revocable
    - Include the cookie in subsequent requests
    - Use the `/api/me` endpoint to verify authentication status

    ## Rate limiting
    - `POST /api/auth/login`: 10 attempts / 15 min per email AND 30 / 15 min per IP
      (successful login resets the email bucket)
    - `POST /api/auth/register`: 10 / 15 min per IP
    - `POST /api/chatbot`: 50 messages / user / UTC day
    - Limited endpoints return 429; auth endpoints include a `Retry-After` header

    ## Conventions
    - **Money:** all amounts are integers in **cents** (e.g. `4999` = $49.99)
    - **IDs:** MongoDB ObjectIds — 24-character hexadecimal strings
    - **Ownership:** on single-resource endpoints, `404` means "not found *or*
      not owned by the requesting user" — the API never reveals whether a
      resource exists for someone else
    - **Dates:** accepted as `YYYY-MM-DD` or full ISO 8601 datetime strings;
      always returned as ISO 8601 datetimes
    - **Errors:** every error response uses the same envelope:
      `{ success: false, error, details? }`
...
      this API

...
servers:
...
paths:
...
  /register:
    get:
      tags: [Client]
      summary: Get registration page
      description: Returns the user registration page.
      responses:
        '200':
          description: Returns the user registration page.
          content:
            text/html:
              schema:
                type: string
        '500':
          $ref: '#/components/responses/InternalServerError'
  /login:
    get:
      tags: [Client]
      summary: Get login page
      description: Returns the user login page.
      responses:
        '200':
          description: Login page rendered successfully.
          content:
            text/html:
              schema:
                type: string
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: |
        Creates a new user account, seeds it with 9 default categories, and
        establishes a session.

        - Emails are normalized to lowercase before storage
        - Passwords are hashed with bcrypt (12 rounds)
        - The session token is stored server-side as a SHA-256 hash
        - Rate limited per IP (see Rate limiting in the API description)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              example1:
                value:
                  name: John Doe
                  email: john@example.com
                  password: password123
      responses:
        '201':
          description: User registered successfully
          headers:
            Set-Cookie:
              description: HTTP-only session cookie
              schema:
                type: string
                example: session_token=Kx8fJ2v...; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: User with this email already exists
        '429':
          description: Too many registrations from this network
          headers:
            Retry-After:
              description: Seconds until the rate-limit window resets
              schema:
                type: integer
                example: 897
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: Too many registrations from this network. Please try again later.
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      description: |
        Authenticates a user and establishes a session.

        - Email matching is case-insensitive
        - Returns the same generic error for unknown emails and wrong
          passwords (no user enumeration)
        - Rate limited per email AND per IP; a successful login resets the
          email bucket (see Rate limiting in the API description)
        - Due recurring income entries are materialized on successful login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              example1:
                value:
                  email: john@example.com
                  password: password123
      responses:
        '200':
          description: Logged in successfully
          headers:
            Set-Cookie:
              description: HTTP-only session cookie
              schema:
                type: string
                example: session_token=Kx8fJ2v...; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: Invalid credentials
        '429':
          description: Too many attempts
          headers:
            Retry-After:
              description: Seconds until the rate-limit window resets
              schema:
                type: integer
                example: 743
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: Too many attempts. Please try again later.
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/expenses/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Expense ObjectId
        schema:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'

    get:
      tags: [Expenses]
      summary: Get an expense
      description: Returns a single expense with its category. A 404 is returned if the expense does not exist or belongs to another user.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Expense retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags: [Expenses]
      summary: Update an expense
      description: |
        Partially updates an expense. Only the provided fields are changed.

        Set `categoryId` to `null` to explicitly remove the expense's category.
        A provided `categoryId` (non-null) must belong to the requesting user.

        Updating an expense also materializes any due recurring income
        entries (see Conventions).
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseUpdateRequest'
            example:
              amount: 6250
              categoryId: null
      responses:
        '200':
          description: Expense updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Expenses]
      summary: Delete an expense
      description: |
        Permanently deletes an expense owned by the authenticated user.
        Deleting also materializes any due recurring income entries.
      security:
        - sessionAuth: []
      responses:
        '204':
          description: Expense deleted successfully (no content)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ------------------------------------------------------------------------

  # Income Routes
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: HTTP-only session cookie obtained from login or registration
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: User's full name
          example: John Doe
        email:
          type: string
          format: email
          description: User's email address (stored lowercase)
          example: john@example.com
      required:
        - id
        - name
        - email
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          description: User's full name
          example: John Doe
        email:
          type: string
          format: email
          description: User's email address (must be unique; normalized to lowercase)
          example: john@example.com
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: |
            User's password (hashed with bcrypt). Max 72 bytes because bcrypt
            silently truncates longer input.
          example: password123
      required:
        - name
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address (case-insensitive)
          example: john@example.com
        password:
          type: string
          description: User's password
          example: password123
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/User'
        message:
          type: string
          example: Logged in successfully
...
    ErrorResponse:
      type: object
      description: |
        Canonical error envelope — every error response in the API uses this
        shape. `details` (Zod issue objects) is present only on validation
        failures.
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Internal Server Error
        details:
          type: array
          description: Present on validation failures (Zod issue objects)
          items:
            $ref: '#/components/schemas/ZodIssue'
        resetAt:
          type: string
          format: date-time
          description: Only on chatbot 429s — when the daily limit resets
      required:
        - success
        - error
    ZodIssue:
      type: object
      description: Validation issue produced by Zod
      properties:
        code:
          type: string
          example: invalid_format
        message:
          type: string
          example: Invalid email address
        path:
          type: array
          items:
            type: string
          example:
            - email
...
    Expense:
      type: object
      properties:
        id:
          type: string
          description: MongoDB ObjectId
          example: 507f1f77bcf86cd799439021
        date:
          type: string
          format: date-time
          description: When the expense occurred
          example: '2025-06-10T00:00:00.000Z'
        name:
          type: string
          example: Weekly groceries
        amount:
          type: integer
          description: Amount in cents
          example: 4999
        categoryId:
          type: string
          nullable: true
          description: Category ObjectId, or null if uncategorized
          example: 507f1f77bcf86cd799439013
        category:
          type: object
          nullable: true
          description: The expense's category, or null if uncategorized
          properties:
            id:
              type: string
              example: 507f1f77bcf86cd799439013
            name:
              type: string
              example: groceries
        userId:
          type: string
          description: Owning user's ObjectId
          example: 507f1f77bcf86cd799439011
      required:
        - id
        - date
        - name
        - amount
        - userId
...
    ExpenseUpdateRequest:
      type: object
      description: All fields optional; only provided fields are updated.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        amount:
          type: integer
          minimum: 1
          description: Amount in cents
        date:
          type: string
          description: 'YYYY-MM-DD or full ISO 8601 datetime'
        categoryId:
          type: string
          nullable: true
          description: 'Category ObjectId, or null to remove the category'
...
  responses:
    Unauthorized:
      description: Authentication required or invalid session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Unauthorized
    BadRequest:
      description: |
        Validation failed (with `details`), a referenced resource is invalid
        (e.g. a categoryId owned by another user), or a query/path parameter
        is malformed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              summary: Validation failure with details
              value:
                success: false
                error: Validation failed
                details:
                  - code: invalid_format
                    message: Invalid email address
                    path:
                      - email
            invalidReference:
              summary: Referenced resource not owned by the user
              value:
                success: false
                error: Category not found
    NotFound:
      description: |
        Resource not found — or it exists but belongs to another user.
        The API does not distinguish between the two cases.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Expense not found
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Internal Server Error